README
This is a Rock, Paper, and Scissors game. Completely text based and created with Python. The user will asked if the user would like to play the game. Upon responding with "Yes" or "No"; the game will either end or start. Once the game starts, the user will be asked to input their choice. Following which, the users choice and the a.i's choice will be compared and depending on the results, the user will be informed if they have won or lost.